Rename SEO expandable section and fix imports
